Refactor code structure by modularizing GitHub API client and types
