+++ src/main.rs
@@ -1,12 +1,14 @@
+mod analyzers;
+mod git;
+mod github;
+mod types;
+
 use anyhow::Result;
-use chrono::{DateTime, Utc};
-use git2::Repository;
+use chrono::Utc;
 use ignore::WalkBuilder;
 use log::{error, info, warn};
 use mime_guess;
 use regex::Regex;
-use reqwest::Client;
-use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
 use std::fs;
 use std::io::Read;
@@ -15,584 +17,19 @@ use std::path::PathBuf;
 use url::Url;
 use walkdir::WalkDir;
 
-// GitHub API response structures
-#[derive(Debug, Serialize, Deserialize, Clone)]
-pub struct GitHubUser {
-    pub login: String,
-    pub id: u64,
-    pub avatar_url: String,
-    pub html_url: String,
-    pub contributions: Option<u32>,
-}
-
-#[derive(Debug, Serialize, Deserialize, Clone)]
-pub struct GitHubLicense {
-    pub key: String,
-    pub name: String,
-    pub spdx_id: Option<String>,
-    pub url: Option<String>,
-}
-
-#[derive(Debug, Serialize, Deserialize, Clone)]
-pub struct GitHubTopic {
-    pub name: String,
-}
-
-#[derive(Debug, Serialize, Deserialize, Clone)]
-pub struct GitHubLanguage {
-    pub name: String,
-    pub bytes: u64,
-    pub percentage: f64,
-}
-
-#[derive(Debug, Serialize, Deserialize, Clone)]
-pub struct GitHubRelease {
-    pub tag_name: String,
-    pub name: Option<String>,
-    pub body: Option<String>,
-    pub draft: bool,
-    pub prerelease: bool,
-    pub created_at: DateTime<Utc>,
-    pub published_at: Option<DateTime<Utc>>,
-    pub author: GitHubUser,
-    pub assets_count: usize,
-}
-
-#[derive(Debug, Serialize, Deserialize, Clone)]
-pub struct GitHubIssue {
-    pub number: u32,
-    pub title: String,
-    pub state: String,
-    pub created_at: DateTime<Utc>,
-    pub updated_at: DateTime<Utc>,
-    pub closed_at: Option<DateTime<Utc>>,
-    pub author: GitHubUser,
-    pub labels: Vec<String>,
-    pub comments: u32,
-}
-
-#[derive(Debug, Serialize, Deserialize, Clone)]
-pub struct GitHubCommit {
-    pub sha: String,
-    pub message: String,
-    pub author: GitHubUser,
-    pub date: DateTime<Utc>,
-    pub additions: u32,
-    pub deletions: u32,
-    pub files_changed: u32,
-}
-
-// Repository metadata from GitHub API
-#[derive(Debug, Serialize, Deserialize, Clone)]
-pub struct RepositoryMetadata {
-    pub id: u64,
-    pub name: String,
-    pub full_name: String,
-    pub description: Option<String>,
-    pub homepage: Option<String>,
-    pub html_url: String,
-    pub clone_url: String,
-    pub ssh_url: String,
-    pub git_url: String,
-    pub owner: GitHubUser,
-    pub private: bool,
-    pub fork: bool,
-    pub archived: bool,
-    pub disabled: bool,
-    pub has_issues: bool,
-    pub has_projects: bool,
-    pub has_wiki: bool,
-    pub has_pages: bool,
-    pub has_downloads: bool,
-    pub has_discussions: bool,
-    pub stargazers_count: u32,
-    pub watchers_count: u32,
-    pub forks_count: u32,
-    pub subscribers_count: Option<u32>,
-    pub network_count: Option<u32>,
-    pub open_issues_count: u32,
-    pub license: Option<GitHubLicense>,
-    pub topics: Vec<String>,
-    pub default_branch: String,
-    pub size: u32, // KB
-    pub language: Option<String>,
-    pub languages: HashMap<String, u64>,
-    pub created_at: DateTime<Utc>,
-    pub updated_at: DateTime<Utc>,
-    pub pushed_at: DateTime<Utc>,
-}
-
-// File analysis structures
-#[derive(Debug, Serialize, Deserialize, Clone)]
-pub struct FileInfo {
-    pub path: PathBuf,
-    pub name: String,
-    pub extension: Option<String>,
-    pub size: u64,
-    pub lines_of_code: Option<u32>,
-    pub blank_lines: Option<u32>,
-    pub comment_lines: Option<u32>,
-    pub language: Option<String>,
-    pub mime_type: Option<String>,
-    pub is_binary: bool,
-    pub is_text: bool,
-    pub encoding: Option<String>,
-    pub hash: String,
-    pub content_preview: Option<String>, // First few lines for analysis
-}
-
-#[derive(Debug, Serialize, Deserialize, Clone)]
-pub struct DirectoryInfo {
-    pub path: PathBuf,
-    pub name: String,
-    pub file_count: u32,
-    pub subdirectory_count: u32,
-    pub total_size: u64,
-    pub files: Vec<FileInfo>,
-    pub subdirectories: Vec<DirectoryInfo>,
-}
-
-// Code analysis structures
-#[derive(Debug, Serialize, Deserialize, Clone)]
-pub struct LanguageStats {
-    pub language: String,
-    pub file_count: u32,
-    pub lines_of_code: u32,
-    pub blank_lines: u32,
-    pub comment_lines: u32,
-    pub total_bytes: u64,
-    pub percentage: f64,
-    pub complexity_score: Option<f64>,
-}
-
-#[derive(Debug, Serialize, Deserialize, Clone)]
-pub struct CodeMetrics {
-    pub total_files: u32,
-    pub total_lines: u32,
-    pub total_loc: u32,
-    pub total_blank_lines: u32,
-    pub total_comment_lines: u32,
-    pub total_size: u64,
-    pub language_stats: HashMap<String, LanguageStats>,
-    pub average_file_size: f64,
-    pub largest_files: Vec<FileInfo>,
-    pub most_complex_files: Vec<FileInfo>,
-}
-
-#[derive(Debug, Serialize, Deserialize, Clone)]
-pub struct ConfigFile {
-    pub path: PathBuf,
-    pub file_type: String, // package.json, Cargo.toml, requirements.txt, etc.
-    pub content: String,
-    pub parsed_dependencies: Option<HashMap<String, String>>,
-    pub scripts: Option<HashMap<String, String>>,
-}
-
-#[derive(Debug, Serialize, Deserialize, Clone)]
-pub struct DocumentationFile {
-    pub path: PathBuf,
-    pub file_type: String, // README, CHANGELOG, LICENSE, etc.
-    pub content: String,
-    pub word_count: u32,
-    pub has_badges: bool,
-    pub has_toc: bool,
-    pub sections: Vec<String>,
-}
-
-// Git analysis structures
-#[derive(Debug, Serialize, Deserialize, Clone)]
-pub struct GitAnalysis {
-    pub total_commits: u32,
-    pub contributors: Vec<GitHubUser>,
-    pub recent_commits: Vec<GitHubCommit>,
-    pub commit_frequency: HashMap<String, u32>, // month -> commit count
-    pub most_active_files: Vec<(String, u32)>,  // file path -> modification count
-    pub branch_count: u32,
-    pub tag_count: u32,
-    pub first_commit_date: Option<DateTime<Utc>>,
-    pub last_commit_date: Option<DateTime<Utc>>,
-}
-
-// Project type detection
-#[derive(Debug, Serialize, Deserialize, Clone)]
-pub struct ProjectInfo {
-    pub primary_language: Option<String>,
-    pub project_type: Vec<String>, // web, cli, library, framework, etc.
-    pub frameworks: Vec<String>,
-    pub build_tools: Vec<String>,
-    pub package_managers: Vec<String>,
-    pub testing_frameworks: Vec<String>,
-    pub ci_cd_tools: Vec<String>,
-    pub deployment_configs: Vec<String>,
-    pub database_technologies: Vec<String>,
-}
-
-// Security and quality analysis
-#[derive(Debug, Serialize, Deserialize, Clone)]
-pub struct SecurityInfo {
-    pub has_security_policy: bool,
-    pub has_dependabot: bool,
-    pub has_codeql: bool,
-    pub vulnerability_alerts: Vec<String>,
-    pub outdated_dependencies: Vec<String>,
-    pub license_compatibility: Vec<String>,
-}
-
-// Comprehensive repository analysis result
-#[derive(Debug, Serialize, Deserialize, Clone)]
-pub struct RepositoryAnalysis {
-    pub url: String,
-    pub analyzed_at: DateTime<Utc>,
-    pub metadata: RepositoryMetadata,
-    pub file_structure: DirectoryInfo,
-    pub code_metrics: CodeMetrics,
-    pub git_analysis: GitAnalysis,
-    pub project_info: ProjectInfo,
-    pub config_files: Vec<ConfigFile>,
-    pub documentation: Vec<DocumentationFile>,
-    pub security_info: SecurityInfo,
-    pub releases: Vec<GitHubRelease>,
-    pub recent_issues: Vec<GitHubIssue>,
-    pub analysis_summary: String,
-    pub ai_insights: Option<String>,
-}
-
-// GitHub API client
-pub struct GitHubClient {
-    client: Client,
-    token: Option<String>,
-    base_url: String,
-}
-
-impl GitHubClient {
-    pub fn new(token: Option<String>) -> Self {
-        Self {
-            client: Client::new(),
-            token,
-            base_url: "https://api.github.com".to_string(),
-        }
-    }
-
-    fn get_auth_headers(&self) -> reqwest::header::HeaderMap {
-        let mut headers = reqwest::header::HeaderMap::new();
-        headers.insert(
-            reqwest::header::USER_AGENT,
-            reqwest::header::HeaderValue::from_static("ai-repo-analyzer-rs/1.0"),
-        );
-
-        if let Some(token) = &self.token {
-            let auth_value = format!("Bearer {}", token);
-            headers.insert(
-                reqwest::header::AUTHORIZATION,
-                reqwest::header::HeaderValue::from_str(&auth_value).unwrap(),
-            );
-        }
-
-        headers
-    }
-
-    pub async fn get_repository_metadata(
-        &self,
-        owner: &str,
-        repo: &str,
-    ) -> Result<RepositoryMetadata> {
-        let url = format!("{}/repos/{}/{}", self.base_url, owner, repo);
-        info!("Fetching repository metadata from: {}", url);
-
-        let response = self
-            .client
-            .get(&url)
-            .headers(self.get_auth_headers())
-            .send()
-            .await?;
-
-        if !response.status().is_success() {
-            anyhow::bail!(
-                "Failed to fetch repository: {} - {}",
-                response.status(),
-                response.text().await?
-            );
-        }
-
-        let repo_data: serde_json::Value = response.json().await?;
-
-        // Fetch additional data
-        let languages = self.get_languages(owner, repo).await.unwrap_or_default();
-        let topics = self.get_topics(owner, repo).await.unwrap_or_default();
-
-        // Convert to our structure
-        let metadata = RepositoryMetadata {
-            id: repo_data["id"].as_u64().unwrap_or(0),
-            name: repo_data["name"].as_str().unwrap_or("").to_string(),
-            full_name: repo_data["full_name"].as_str().unwrap_or("").to_string(),
-            description: repo_data["description"].as_str().map(|s| s.to_string()),
-            homepage: repo_data["homepage"].as_str().map(|s| s.to_string()),
-            html_url: repo_data["html_url"].as_str().unwrap_or("").to_string(),
-            clone_url: repo_data["clone_url"].as_str().unwrap_or("").to_string(),
-            ssh_url: repo_data["ssh_url"].as_str().unwrap_or("").to_string(),
-            git_url: repo_data["git_url"].as_str().unwrap_or("").to_string(),
-            owner: GitHubUser {
-                login: repo_data["owner"]["login"]
-                    .as_str()
-                    .unwrap_or("")
-                    .to_string(),
-                id: repo_data["owner"]["id"].as_u64().unwrap_or(0),
-                avatar_url: repo_data["owner"]["avatar_url"]
-                    .as_str()
-                    .unwrap_or("")
-                    .to_string(),
-                html_url: repo_data["owner"]["html_url"]
-                    .as_str()
-                    .unwrap_or("")
-                    .to_string(),
-                contributions: None,
-            },
-            private: repo_data["private"].as_bool().unwrap_or(false),
-            fork: repo_data["fork"].as_bool().unwrap_or(false),
-            archived: repo_data["archived"].as_bool().unwrap_or(false),
-            disabled: repo_data["disabled"].as_bool().unwrap_or(false),
-            has_issues: repo_data["has_issues"].as_bool().unwrap_or(false),
-            has_projects: repo_data["has_projects"].as_bool().unwrap_or(false),
-            has_wiki: repo_data["has_wiki"].as_bool().unwrap_or(false),
-            has_pages: repo_data["has_pages"].as_bool().unwrap_or(false),
-            has_downloads: repo_data["has_downloads"].as_bool().unwrap_or(false),
-            has_discussions: repo_data["has_discussions"].as_bool().unwrap_or(false),
-            stargazers_count: repo_data["stargazers_count"].as_u64().unwrap_or(0) as u32,
-            watchers_count: repo_data["watchers_count"].as_u64().unwrap_or(0) as u32,
-            forks_count: repo_data["forks_count"].as_u64().unwrap_or(0) as u32,
-            subscribers_count: repo_data["subscribers_count"].as_u64().map(|x| x as u32),
-            network_count: repo_data["network_count"].as_u64().map(|x| x as u32),
-            open_issues_count: repo_data["open_issues_count"].as_u64().unwrap_or(0) as u32,
-            license: repo_data["license"]
-                .as_object()
-                .map(|license| GitHubLicense {
-                    key: license["key"].as_str().unwrap_or("").to_string(),
-                    name: license["name"].as_str().unwrap_or("").to_string(),
-                    spdx_id: license["spdx_id"].as_str().map(|s| s.to_string()),
-                    url: license["url"].as_str().map(|s| s.to_string()),
-                }),
-            topics,
-            default_branch: repo_data["default_branch"]
-                .as_str()
-                .unwrap_or("main")
-                .to_string(),
-            size: repo_data["size"].as_u64().unwrap_or(0) as u32,
-            language: repo_data["language"].as_str().map(|s| s.to_string()),
-            languages,
-            created_at: chrono::DateTime::parse_from_rfc3339(
-                repo_data["created_at"]
-                    .as_str()
-                    .unwrap_or("1970-01-01T00:00:00Z"),
-            )
-            .unwrap()
-            .with_timezone(&Utc),
-            updated_at: chrono::DateTime::parse_from_rfc3339(
-                repo_data["updated_at"]
-                    .as_str()
-                    .unwrap_or("1970-01-01T00:00:00Z"),
-            )
-            .unwrap()
-            .with_timezone(&Utc),
-            pushed_at: chrono::DateTime::parse_from_rfc3339(
-                repo_data["pushed_at"]
-                    .as_str()
-                    .unwrap_or("1970-01-01T00:00:00Z"),
-            )
-            .unwrap()
-            .with_timezone(&Utc),
-        };
-
-        Ok(metadata)
-    }
-
-    pub async fn get_languages(&self, owner: &str, repo: &str) -> Result<HashMap<String, u64>> {
-        let url = format!("{}/repos/{}/{}/languages", self.base_url, owner, repo);
-
-        let response = self
-            .client
-            .get(&url)
-            .headers(self.get_auth_headers())
-            .send()
-            .await?;
-
-        if response.status().is_success() {
-            let languages: HashMap<String, u64> = response.json().await?;
-            Ok(languages)
-        } else {
-            Ok(HashMap::new())
-        }
-    }
-
-    pub async fn get_topics(&self, owner: &str, repo: &str) -> Result<Vec<String>> {
-        let url = format!("{}/repos/{}/{}/topics", self.base_url, owner, repo);
-
-        let response = self
-            .client
-            .get(&url)
-            .headers(self.get_auth_headers())
-            .send()
-            .await?;
-
-        if response.status().is_success() {
-            let data: serde_json::Value = response.json().await?;
-            let topics = data["names"]
-                .as_array()
-                .unwrap_or(&Vec::new())
-                .iter()
-                .filter_map(|v| v.as_str())
-                .map(|s| s.to_string())
-                .collect();
-            Ok(topics)
-        } else {
-            Ok(Vec::new())
-        }
-    }
-
-    pub async fn get_contributors(&self, owner: &str, repo: &str) -> Result<Vec<GitHubUser>> {
-        let url = format!("{}/repos/{}/{}/contributors", self.base_url, owner, repo);
-
-        let response = self
-            .client
-            .get(&url)
-            .headers(self.get_auth_headers())
-            .send()
-            .await?;
-
-        if response.status().is_success() {
-            let contributors: Vec<serde_json::Value> = response.json().await?;
-            let users = contributors
-                .into_iter()
-                .map(|c| GitHubUser {
-                    login: c["login"].as_str().unwrap_or("").to_string(),
-                    id: c["id"].as_u64().unwrap_or(0),
-                    avatar_url: c["avatar_url"].as_str().unwrap_or("").to_string(),
-                    html_url: c["html_url"].as_str().unwrap_or("").to_string(),
-                    contributions: c["contributions"].as_u64().map(|x| x as u32),
-                })
-                .collect();
-            Ok(users)
-        } else {
-            Ok(Vec::new())
-        }
-    }
-
-    pub async fn get_releases(
-        &self,
-        owner: &str,
-        repo: &str,
-        limit: usize,
-    ) -> Result<Vec<GitHubRelease>> {
-        let url = format!(
-            "{}/repos/{}/{}/releases?per_page={}",
-            self.base_url, owner, repo, limit
-        );
-
-        let response = self
-            .client
-            .get(&url)
-            .headers(self.get_auth_headers())
-            .send()
-            .await?;
-
-        if response.status().is_success() {
-            let releases: Vec<serde_json::Value> = response.json().await?;
-            let parsed_releases = releases
-                .into_iter()
-                .map(|r| GitHubRelease {
-                    tag_name: r["tag_name"].as_str().unwrap_or("").to_string(),
-                    name: r["name"].as_str().map(|s| s.to_string()),
-                    body: r["body"].as_str().map(|s| s.to_string()),
-                    draft: r["draft"].as_bool().unwrap_or(false),
-                    prerelease: r["prerelease"].as_bool().unwrap_or(false),
-                    created_at: chrono::DateTime::parse_from_rfc3339(
-                        r["created_at"].as_str().unwrap_or("1970-01-01T00:00:00Z"),
-                    )
-                    .unwrap()
-                    .with_timezone(&Utc),
-                    published_at: r["published_at"]
-                        .as_str()
-                        .and_then(|s| chrono::DateTime::parse_from_rfc3339(s).ok())
-                        .map(|dt| dt.with_timezone(&Utc)),
-                    author: GitHubUser {
-                        login: r["author"]["login"].as_str().unwrap_or("").to_string(),
-                        id: r["author"]["id"].as_u64().unwrap_or(0),
-                        avatar_url: r["author"]["avatar_url"].as_str().unwrap_or("").to_string(),
-                        html_url: r["author"]["html_url"].as_str().unwrap_or("").to_string(),
-                        contributions: None,
-                    },
-                    assets_count: r["assets"].as_array().map(|a| a.len()).unwrap_or(0),
-                })
-                .collect();
-            Ok(parsed_releases)
-        } else {
-            Ok(Vec::new())
-        }
-    }
-
-    pub async fn get_recent_issues(
-        &self,
-        owner: &str,
-        repo: &str,
-        limit: usize,
-    ) -> Result<Vec<GitHubIssue>> {
-        let url = format!(
-            "{}/repos/{}/{}/issues?state=all&per_page={}&sort=updated",
-            self.base_url, owner, repo, limit
-        );
-
-        let response = self
-            .client
-            .get(&url)
-            .headers(self.get_auth_headers())
-            .send()
-            .await?;
-
-        if response.status().is_success() {
-            let issues: Vec<serde_json::Value> = response.json().await?;
-            let parsed_issues = issues
-                .into_iter()
-                .filter(|i| i["pull_request"].is_null()) // Filter out pull requests
-                .map(|i| GitHubIssue {
-                    number: i["number"].as_u64().unwrap_or(0) as u32,
-                    title: i["title"].as_str().unwrap_or("").to_string(),
-                    state: i["state"].as_str().unwrap_or("").to_string(),
-                    created_at: chrono::DateTime::parse_from_rfc3339(
-                        i["created_at"].as_str().unwrap_or("1970-01-01T00:00:00Z"),
-                    )
-                    .unwrap()
-                    .with_timezone(&Utc),
-                    updated_at: chrono::DateTime::parse_from_rfc3339(
-                        i["updated_at"].as_str().unwrap_or("1970-01-01T00:00:00Z"),
-                    )
-                    .unwrap()
-                    .with_timezone(&Utc),
-                    closed_at: i["closed_at"]
-                        .as_str()
-                        .and_then(|s| chrono::DateTime::parse_from_rfc3339(s).ok())
-                        .map(|dt| dt.with_timezone(&Utc)),
-                    author: GitHubUser {
-                        login: i["user"]["login"].as_str().unwrap_or("").to_string(),
-                        id: i["user"]["id"].as_u64().unwrap_or(0),
-                        avatar_url: i["user"]["avatar_url"].as_str().unwrap_or("").to_string(),
-                        html_url: i["user"]["html_url"].as_str().unwrap_or("").to_string(),
-                        contributions: None,
-                    },
-                    labels: i["labels"]
-                        .as_array()
-                        .unwrap_or(&Vec::new())
-                        .iter()
-                        .filter_map(|l| l["name"].as_str())
-                        .map(|s| s.to_string())
-                        .collect(),
-                    comments: i["comments"].as_u64().unwrap_or(0) as u32,
-                })
-                .collect();
-            Ok(parsed_issues)
-        } else {
-            Ok(Vec::new())
-        }
-    }
-}
+use crate::git::GitManager;
+use crate::github::GitHubClient;
+use crate::types::CodeMetrics;
+use crate::types::ConfigFile;
+use crate::types::DirectoryInfo;
+use crate::types::DocumentationFile;
+use crate::types::FileInfo;
+use crate::types::GitAnalysis;
+use crate::types::LanguageStats;
+use crate::types::ProjectInfo;
+use crate::types::RepositoryAnalysis;
+use crate::types::RepositoryMetadata;
+use crate::types::SecurityInfo;
 
 // Utility function to parse GitHub URL
 pub fn parse_github_url(url: &str) -> Result<(String, String)> {
@@ -615,170 +52,6 @@ pub fn parse_github_url(url: &str) -> Result<(String, String)> {
     let repo = path_segments[1].trim_end_matches(".git").to_string();
 
     Ok((owner, repo))
-}
-
-// Git repository manager
-pub struct GitManager {
-    work_dir: PathBuf,
-}
-
-impl GitManager {
-    pub fn new(work_dir: Option<PathBuf>) -> Self {
-        let work_dir = work_dir.unwrap_or_else(|| std::env::temp_dir().join("ai-repo-analyzer"));
-
-        // Create work directory if it doesn't exist
-        if !work_dir.exists() {
-            std::fs::create_dir_all(&work_dir).unwrap_or_else(|e| {
-                warn!("Failed to create work directory: {}", e);
-            });
-        }
-
-        Self { work_dir }
-    }
-
-    pub async fn clone_or_update_repository(
-        &self,
-        clone_url: &str,
-        repo_name: &str,
-    ) -> Result<PathBuf> {
-        let repo_path = self.work_dir.join(repo_name);
-
-        // Remove existing directory if it exists
-        if repo_path.exists() {
-            info!("Removing existing repository directory: {:?}", repo_path);
-            fs::remove_dir_all(&repo_path)?;
-        }
-
-        info!("Cloning repository from {} to {:?}", clone_url, repo_path);
-
-        // Clone the repository
-        let _repo = Repository::clone(clone_url, &repo_path)
-            .map_err(|e| anyhow::anyhow!("Failed to clone repository: {}", e))?;
-
-        info!("Successfully cloned repository to {:?}", repo_path);
-        Ok(repo_path)
-    }
-
-    pub fn analyze_git_history(&self, repo_path: &Path) -> Result<GitAnalysis> {
-        let repo = Repository::open(repo_path)?;
-
-        // Get all commits
-        let mut revwalk = repo.revwalk()?;
-        revwalk.push_head()?;
-        revwalk.set_sorting(git2::Sort::TIME)?;
-
-        let mut total_commits = 0;
-        let mut contributors: HashMap<String, GitHubUser> = HashMap::new();
-        let mut recent_commits = Vec::new();
-        let mut commit_frequency: HashMap<String, u32> = HashMap::new();
-        let mut file_modifications: HashMap<String, u32> = HashMap::new();
-        let mut first_commit_date: Option<DateTime<Utc>> = None;
-        let mut last_commit_date: Option<DateTime<Utc>> = None;
-
-        for (index, oid) in revwalk.enumerate() {
-            if index >= 1000 {
-                // Limit to first 1000 commits for performance
-                break;
-            }
-
-            let oid = oid?;
-            let commit = repo.find_commit(oid)?;
-            total_commits += 1;
-
-            let commit_time = DateTime::from_timestamp(commit.time().seconds(), 0)
-                .unwrap_or_else(|| Utc::now())
-                .with_timezone(&Utc);
-
-            if first_commit_date.is_none() {
-                first_commit_date = Some(commit_time);
-            }
-            last_commit_date = Some(commit_time);
-
-            // Track commit frequency by month
-            let month_key = commit_time.format("%Y-%m").to_string();
-            *commit_frequency.entry(month_key).or_insert(0) += 1;
-
-            // Track contributors
-            let author = commit.author();
-            if let (Some(name), Some(email)) = (author.name(), author.email()) {
-                let key = format!("{}:{}", name, email);
-                contributors
-                    .entry(key.clone())
-                    .or_insert_with(|| GitHubUser {
-                        login: name.to_string(),
-                        id: 0, // We don't have GitHub ID from git history
-                        avatar_url: String::new(),
-                        html_url: String::new(),
-                        contributions: Some(0),
-                    });
-                if let Some(user) = contributors.get_mut(&key) {
-                    user.contributions = Some(user.contributions.unwrap_or(0) + 1);
-                }
-            }
-
-            // Store recent commits (first 50)
-            if recent_commits.len() < 50 {
-                let git_commit = GitHubCommit {
-                    sha: format!("{}", oid),
-                    message: commit.message().unwrap_or("").to_string(),
-                    author: GitHubUser {
-                        login: author.name().unwrap_or("Unknown").to_string(),
-                        id: 0,
-                        avatar_url: String::new(),
-                        html_url: String::new(),
-                        contributions: None,
-                    },
-                    date: commit_time,
-                    additions: 0, // Git2 doesn't provide easy access to diff stats
-                    deletions: 0,
-                    files_changed: 0,
-                };
-                recent_commits.push(git_commit);
-            }
-
-            // Track file modifications (simplified)
-            if let Ok(tree) = commit.tree() {
-                let mut file_count = 0;
-                tree.walk(git2::TreeWalkMode::PreOrder, |_root, entry| {
-                    if let Some(name) = entry.name() {
-                        *file_modifications.entry(name.to_string()).or_insert(0) += 1;
-                        file_count += 1;
-                    }
-                    if file_count > 100 {
-                        // Limit file tracking for performance
-                        git2::TreeWalkResult::Abort
-                    } else {
-                        git2::TreeWalkResult::Ok
-                    }
-                })?;
-            }
-        }
-
-        // Get most active files
-        let mut most_active_files: Vec<_> = file_modifications.into_iter().collect();
-        most_active_files.sort_by(|a, b| b.1.cmp(&a.1));
-        most_active_files.truncate(20);
-
-        // Count branches and tags
-        let branches = repo.branches(Some(git2::BranchType::Local))?;
-        let branch_count = branches.count() as u32;
-
-        let tag_count = repo.tag_names(None)?.len() as u32;
-
-        let git_analysis = GitAnalysis {
-            total_commits,
-            contributors: contributors.into_values().collect(),
-            recent_commits,
-            commit_frequency,
-            most_active_files,
-            branch_count,
-            tag_count,
-            first_commit_date,
-            last_commit_date,
-        };
-
-        Ok(git_analysis)
-    }
 }
 
 // File system analyzer
